refactor: 트랜잭션 관련 로직 분리를 위한 UserService 추상화
